requirements.txt
aiosqlite==0.17.0
anyio==3.7.1
asyncmy==0.2.8
click==8.1.6
exceptiongroup==1.1.2
h11==0.14.0
httptools==0.6.0
idna==3.4
iso8601==1.1.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
sniffio==1.3.0
tortoise-orm==0.20.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
